[python] refactor: Packaing modules - Package modules to lassl
Refs: #75
