+++ pretrain_language_model.py
@@ -2,6 +2,7 @@ import logging
 import os
 from argparse import ArgumentParser
 
+from datasets import Dataset
 from omegaconf import OmegaConf
 from transformers import (
     CONFIG_MAPPING,
@@ -13,8 +14,7 @@ from transformers import (
 )
 from transformers.trainer_utils import get_last_checkpoint
 
-from datasets import Dataset
-from src.collators import (
+from lassl.collators import (
     DataCollatorForAlbert,
     DataCollatorForBert,
     DataCollatorForGpt2,
