+++ requirements.txt
@@ -1,8 +1,56 @@
+aiohttp==3.8.1
+aiosignal==1.2.0
+antlr4-python3-runtime==4.8
+async-timeout==4.0.2
+attrs==21.4.0
+black==21.12b0
+certifi==2021.10.8
+charset-normalizer==2.0.12
+click==8.0.4
 datasets==1.16.1
+dill==0.3.4
+docopt==0.6.2
+filelock==3.6.0
+frozenlist==1.3.0
+fsspec==2022.2.0
+huggingface-hub==0.4.0
+idna==3.3
+iniconfig==1.1.1
+isort==5.10.1
+joblib==1.1.0
+langid==1.1.6
+multidict==6.0.2
+multiprocess==0.70.12.2
+mypy-extensions==0.4.3
 ninja==1.10.2.3
 numpy==1.21.1
+omegaconf==2.1.1
+packaging==21.3
+pandas==1.4.1
+pathspec==0.9.0
+pipreqs==0.4.11
+platformdirs==2.5.1
+pluggy==1.0.0
+py==1.11.0
 pyarrow==6.0.1
 pybind11==2.9.1
+pyparsing==3.0.7
+pytest==7.1.0
+python-dateutil==2.8.2
+pytz==2021.3
+PyYAML==6.0
+regex==2022.3.15
+requests==2.27.1
+responses==0.18.0
+sacremoses==0.0.49
+six==1.16.0
+tokenizers==0.10.3
+tomli==1.2.3
 torch==1.10.0
 tqdm==4.62.3
 transformers==4.13.0
+typing_extensions==4.1.1
+urllib3==1.26.9
+xxhash==3.0.0
+yarg==0.1.9
+yarl==1.7.2
